Provide a zip of the docs + library on deploy
Moves the docs/ into the dist/ directory.
Adds the compress task, to zip up everything in the dist/ directory.
The zip will be added to the git repository, in place of the dist/ and
docs/ directories.

Also simplifies the watch task to call "deploy" instead of a list of jobs