Remove axis lines from index page
Removes the css which was overriding everything on the axis
We now specify tickSize(0) to remove the axis lines.
We also remove the unused monthNames in the connect example.

The latest update to the gitignore means that we ignore everything in
the dist folder, except for the insight.min.js file.