Create Legend
Adds a legend to the chart, drawn from each series' name and colour.
Position of the legend is currently hardcoded to be the top-right(isn)
corner.

The sizing of the chart elements was being handled by the init() and
resizeChart(), the latter of which is called by draw(). Instead, all
resizing is now handled solely by resizeChart().

I also found that insight.Constants.Chart no longer exists, as it is now
the PlotArea, so i've fixed this up too.