Fix Automargin
The autosizing margins were not being applied at the right time (only in
init instead of draw). The labels on the margin can change as the chart
is dragged around, so it needs to be on draw.

It was also not taking the tick size, tick padding, and axis label size
into account when calculating the margins.