Fix zoom-tooltip configuration issues
Zoom needs to be set up after the axes have first been drawn, but the
zoom pane needs to be the first element in the dom. We now set up the
axes first, but add the zoom pane before the first-child of the dom,
allowing both tooltips and zooming to work simultaneously.