Tweaks to charts
Removes the rotation on a few of the axis tick labels, as this makes
them harder to skim read.

Makes the genre chart a row chart, instead of columns, to better show
the long-labelled key axis without tick rotation.

Also adds the x-axis back to this chart, so that data can be read off
without the tooltip.