Fix issue with points on multiple line series
The line series were only selecting all circles on the chart, then
comparing against their own data. Now, we use a classname for the line
points, which is unique for the line (assuming names are unique).

This lets us have 2 line series on a chart, and configure whether they
show or hide points independent of each other.