Fixed exceptions caused by angular $apply within $apply
The ChartGroup.filterChanged event is called either as part of clicking
on the chart or the the hyperlink, the latter uses Angular bindings.

In order for Angular to detect that the scope has changed, $apply must
be called. However you can not call $apply from an Angular bound event,
so a conditional check needed to be introduced before calling $apply.