Small code tidy-ups
Removes some of the more verbose parts of the snippet code, as it was
making the tooltip excessively large.

Also changes the </br> tag in the bubble chart on-chart-tooltip for a
"\n", as this was being interpreted by the show-code-tooltip as HTML to
apply, rather than as something to be in text.